Update masahat mosalasl.py

+++ masahat mosalasl.py
@@ -1,5 +1,5 @@
-ghaede=int(input("enter ghaede:"))
-ertefae=int(input("enter ertefae:"))
+ghaede=float(input("enter ghaede:"))
+ertefae=float(input("enter ertefae:"))
 masahat=(ghaede*ertefae)/2
 print("masahat mosalas:",masahat)
